refactor: 랭킹 도메인 기능 리팩터링

+++ src/test/java/dasi/typing/domain/typing/TypingRepositoryTest.java
@@ -254,7 +254,7 @@ class TypingRepositoryTest {
   private Typing createTyping(int cpm, int acc, Member member, Phrase phrase) {
     return Typing.builder()
         .cpm(cpm)
-        .acc(acc)
+        .acc((double) acc)
         .wpm(0)
         .maxCpm(0)
         .member(member)
